+++ wee/webscraper/src/app/app.module.ts
@@ -1,8 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ScraperModule } from '../scraper/scraper.module';
-
+import { ScreenshotModule } from '../homepage-screenshot/screenshot.modules';
 @Module({
-  imports: [ScraperModule],
+  imports: [ScraperModule, ScreenshotModule],
 })
 export class AppModule {}
-
