--- wee/webscraper/src/homepage-screenshot/screenshot.modules.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ScreenshotService } from './screenshot.service';
-import { ScreenshotController } from './screenshot.controller';
-
-@Module({
-  providers: [ScreenshotService],
-  controllers: [ScreenshotController],
-})
-export class ScreenshotModule {}
